Создание бота

+++ taski/app/crud/task.py
@@ -1,19 +1,11 @@
-from datetime import timedelta, datetime
+from typing import Optional
 
+from fastapi import HTTPException
+from fastapi.encoders import jsonable_encoder
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
-from fastapi.encoders import jsonable_encoder
 
 from app.models import Task, User, Tag
-
-
-def format_timediff(timediff: timedelta) -> str:
-    total_seconds = int(timediff.total_seconds())
-    days, remainder = divmod(total_seconds, 86400)
-    hours, remainder = divmod(remainder, 3600)
-    minutes, remainder = divmod(remainder, 60)
-    seconds, microseconds = divmod(remainder, 1)
-    return f"{days} days, {hours}:{minutes}:{seconds}.{microseconds:06d}"
 
 
 class CRUDTask:
@@ -56,8 +48,7 @@ class CRUDTask:
     ):
         """Создание объекта в БД."""
         obj_in_data = obj_in.dict()
-        if user is not None:
-            obj_in_data['user_id'] = user.id
+        obj_in_data['user_id'] = user.id
         db_obj = self.model(**obj_in_data)
         session.add(db_obj)
         await session.commit()
@@ -68,40 +59,71 @@ class CRUDTask:
             self,
             db_obj,
             obj_in,
-            session: AsyncSession
+            session: AsyncSession,
+            user: User
     ):
         """Обновление данных объекта."""
+        if db_obj.user_id != user.id:
+            raise HTTPException(
+                status_code=403,
+                detail="Not authorized to update this task"
+            )
+
         obj_data = jsonable_encoder(db_obj)
         update_data = obj_in.dict(exclude_unset=True)
 
         for field in obj_data:
             if field in update_data:
                 setattr(db_obj, field, update_data[field])
-        db_obj.update_date = datetime.now()
         session.add(db_obj)
         await session.commit()
         await session.refresh(db_obj)
         return db_obj
 
     async def remove(
-            self,
-            db_obj,
-            session: AsyncSession
+        self,
+        db_obj,
+        session: AsyncSession,
+        user: User
     ):
         """Удаление объекта."""
+        if db_obj.user_id != user.id:
+            raise HTTPException(
+                status_code=403,
+                detail="Not authorized to delete this task"
+            )
         await session.delete(db_obj)
         await session.commit()
         return db_obj
 
     async def get_tasks_by_tag(
         tag: str,
-        session: AsyncSession
+        session: AsyncSession,
+        user: User
     ) -> list[Task]:
         """Получение задач по тегу."""
         result = await session.execute(
-            select(Task).join(Task.tags).where(Tag.title == tag)
+            select(Task).join(Task.tags).where(
+                Tag.title == tag,
+                Task.user_id == user.id
+            )
         )
         return result.scalars().all()
+
+    async def get_task_id_by_name(
+        self,
+        task_name: str,
+        session: AsyncSession,
+        user: User
+    ) -> Optional[int]:
+        """Получение объекта по имени."""
+        db_task_id = await session.execute(
+            select(Task.id).where(
+                Task.name == task_name,
+                Task.user_id == user.id
+            )
+        )
+        return db_task_id.scalars().first()
 
 
 task_crud = CRUDTask(Task)
